confirmedAt перенесён в part из всех частей
