+++ src/Entity/Document.php
@@ -49,13 +49,6 @@ class Document
     private $event;
 
     /**
-     * @var DateTimeImmutable
-     *
-     * @ORM\Column(type="datetime_immutable", nullable=true)
-     */
-    private $confirmedAt;
-
-    /**
      * @return int|null
      */
     public function getId(): ?int
@@ -139,26 +132,6 @@ class Document
     public function setEvent(Event $event): self
     {
         $this->event = $event;
-
-        return $this;
-    }
-
-    /**
-     * @return DateTimeImmutable|null
-     */
-    public function getConfirmedAt(): ?DateTimeImmutable
-    {
-        return $this->confirmedAt;
-    }
-
-    /**
-     * @param DateTimeImmutable|null $confirmedAt
-     *
-     * @return $this
-     */
-    public function setConfirmedAt(?DateTimeImmutable $confirmedAt): self
-    {
-        $this->confirmedAt = $confirmedAt;
 
         return $this;
     }
