+++ src/Entity/Transport.php
@@ -51,13 +51,6 @@ class Transport
     private $event;
 
     /**
-     * @var DateTimeImmutable
-     *
-     * @ORM\Column(type="datetime_immutable", nullable=true)
-     */
-    private $confirmedAt;
-
-    /**
      * @return int|null
      */
     public function getId(): ?int
@@ -143,26 +136,6 @@ class Transport
     public function setEvent(Event $event): self
     {
         $this->event = $event;
-
-        return $this;
-    }
-
-    /**
-     * @return DateTimeImmutable|null
-     */
-    public function getConfirmedAt(): ?DateTimeImmutable
-    {
-        return $this->confirmedAt;
-    }
-
-    /**
-     * @param DateTimeImmutable $confirmedAt
-     *
-     * @return $this
-     */
-    public function setConfirmedAt(DateTimeImmutable $confirmedAt): self
-    {
-        $this->confirmedAt = $confirmedAt;
 
         return $this;
     }
